requirements.txt
pycryptodome
python-dotenv
requests
websocket-client
